requirements.txt
PyQt5>=5.15.0
PyQt5-stubs>=5.15.0
pytesseract>=0.3.0
Pillow>=8.0.0
openai>=1.0.0
typing-extensions>=4.0.0
types-Pillow>=8.0.0
types-pytesseract>=0.3.0
mypy>=1.0.0
types-setuptools>=67.0.0 